+++ myRenderer3/myRenderer3.vcxproj
@@ -137,7 +137,7 @@
     <ClCompile Include="logger.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="processArgs.cpp" />
-    <ClCompile Include="profiler.cpp" />
+    <ClCompile Include="profileManager.cpp" />
     <ClCompile Include="runMainLoop.cpp" />
     <ClCompile Include="stack.cpp" />
     <ClCompile Include="test.cpp" />
@@ -154,7 +154,7 @@
     <ClInclude Include="getKeyInput.h" />
     <ClInclude Include="logger.h" />
     <ClInclude Include="processArgs.h" />
-    <ClInclude Include="profiler.h" />
+    <ClInclude Include="profileManager.h" />
     <ClInclude Include="runMainLoop.h" />
     <ClInclude Include="stack.h" />
     <ClInclude Include="test.h" />
